requirements.txt
faster-whisper>=1.1.0

# Existing transcription / dataset pipeline dependencies
openai-whisper
torch
yt-dlp

# Note:
# - ffmpeg is a system dependency (required for audio decoding/extraction).
# - torch may need a platform/GPU-specific installation command:
#   https://pytorch.org/get-started/locally/
